feature: add support for fs_err
