+++ rust-toolchain.toml
@@ -1,0 +1,2 @@
+[toolchain]
+channel = 'nightly'
